Some class Code cleaned

+++ src/StrongholdCrusader/GameObjects/Humans/Soldier.java
@@ -59,7 +59,7 @@ public class Soldier extends Human {
         imageView.setLayoutY(40);
         button.setLayoutX(200);
         button.setLayoutY(50);
-        ProgressBar health = new ProgressBar(this.health/100);
+        ProgressBar health = new ProgressBar((double)this.health/100);
         health.setLayoutX(Settings.MENUS_ANCHORPANE_WIDTH - 150);
         health.setStyle("-fx-accent: #96ff4c;");
         health.setLayoutY(20);
